EZP-24711: Moved RoleData to Role NS
