+++ tests/RepositoryForms/Data/Mapper/RoleMapperTest.php
@@ -11,7 +11,7 @@ namespace EzSystems\RepositoryForms\Tests\Data\Mapper;
 use eZ\Publish\Core\Repository\Values\User\Role;
 use eZ\Publish\Core\Repository\Values\User\RoleDraft;
 use EzSystems\RepositoryForms\Data\Mapper\RoleMapper;
-use EzSystems\RepositoryForms\Data\RoleData;
+use EzSystems\RepositoryForms\Data\Role\RoleData;
 use PHPUnit_Framework_TestCase;
 
 class RoleMapperTest extends PHPUnit_Framework_TestCase
